requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Experiment Tracking
mlflow>=2.8.0

# Data Versioning
dvc>=3.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Monitoring
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
